Show donor diversity in search results
Redmine #4543
