+++ src/encoded/static/components/dataset.js
@@ -10,6 +10,7 @@ import { FetchedItems } from './fetched';
 import { AuditIndicators, AuditDetail, AuditMixin } from './audit';
 import { StatusLabel } from './statuslabel';
 import { PubReferenceList } from './reference';
+import { donorDiversity } from './objectutils';
 import { softwareVersionList } from './software';
 import { SortTablePanel, SortTable } from './sorttable';
 import { ProjectBadge } from './image';
@@ -948,67 +949,6 @@ function displayPossibleControls(item) {
         );
     }
     return null;
-}
-
-
-// Given a dataset (for now, only ReferenceEpigenome), return the donor diversity of that dataset.
-function donorDiversity(dataset) {
-    let diversity = 'none';
-
-    if (dataset.related_datasets && dataset.related_datasets.length) {
-        // Get all non-deleted related experiments; empty array if none.
-        const experiments = dataset.related_datasets.filter(experiment => experiment.status !== 'deleted');
-
-        // From list list of non-deleted experiments, get all non-deleted replicates into one
-        // array.
-        if (experiments.length) {
-            // Make an array of replicate arrays, one replicate array per experiment. Only include
-            // non-deleted replicates.
-            const replicatesByExperiment = experiments.map(experiment => (
-                (experiment.replicates && experiment.replicates.length) ?
-                    experiment.replicates.filter(replicate => replicate.status !== 'deleted')
-                : []),
-            );
-
-            // Merge all replicate arrays into one non-deleted replicate array.
-            const replicates = replicatesByExperiment.reduce((replicateCollection, replicatesForExperiment) => replicateCollection.concat(replicatesForExperiment), []);
-
-            // Look at the donors in each replicate's biosample. If we see at least two different
-            // donors, we know we have a composite. If only one unique donor after examining all
-            // donors, we have a single. "None" if no donors found in all replicates.
-            if (replicates.length) {
-                const donorAtIdCollection = [];
-                replicates.every((replicate) => {
-                    if (replicate.library && replicate.library.status !== 'deleted' &&
-                            replicate.library.biosample && replicate.library.biosample.status !== 'deleted' &&
-                            replicate.library.biosample.donor && replicate.library.biosample.donor.status !== 'deleted') {
-                        const donorAtId = replicate.library.biosample.donor['@id'];
-
-                        // If we haven't yet seen this donor @id, add it to our collection
-                        if (donorAtIdCollection.indexOf(donorAtId) === -1) {
-                            donorAtIdCollection.push(donorAtId);
-                        }
-
-                        // If we have two, we know have a composite, and we can exit the loop by
-                        // returning false, which makes the replicates.every function end.
-                        return donorAtIdCollection.length !== 2;
-                    }
-
-                    // No donor to examine in this replicate. Keep the `every` loop going.
-                    return true;
-                });
-
-                // Now determine the donor diversity.
-                if (donorAtIdCollection.length > 1) {
-                    diversity = 'composite';
-                } else if (donorAtIdCollection.length === 1) {
-                    diversity = 'single';
-                } // Else keep its original value of 'none'.
-            }
-        }
-    }
-    console.log('DIVERSITY: %s', diversity);
-    return diversity;
 }
